backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
openai==1.11.1
python-whois==0.8.0
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3 